swagger: Dynamically Generates Documentation from a Swagger-Compliant API

Swagger is a collection of HTML, JavaScript, and CSS assets that dynamically generate beautiful documentation from a Swagger-compliant API.

Details

The main purpose of this package is to enable package authors to create APIs that are compatible with https://swagger.io/ and https://www.openapis.org/.

To learn more about Swagger visit: https://swagger.io/swagger-ui/

See also

Author

Maintainer: Bruno Tremblay cran@neoxone.com [contributor]

Authors:

Barret Schloerke barret@rstudio.com (ORCID)

Javier Luraschi javier@rstudio.com

SmartBear Software [copyright holder]

Other contributors:

RStudio [copyright holder]